# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "Authenticate user with username and password",
                "consumes": [
                    "application/json",
                    "multipart/form-data",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "$ref": "#/definitions/controllers.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to generate token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Invalidate user's JWT token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logout user",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "schema": {
                            "$ref": "#/definitions/controllers.LogoutResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid authorization header",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get authenticated user's profile information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get user profile",
                "responses": {
                    "200": {
                        "description": "User profile",
                        "schema": {
                            "$ref": "#/definitions/controllers.ProfileResponse"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to get user profile",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/refresh-token": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generate a new JWT token using current valid token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh JWT token",
                "responses": {
                    "200": {
                        "description": "Token refreshed",
                        "schema": {
                            "$ref": "#/definitions/controllers.RefreshTokenResponse"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to generate token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/universal-login": {
            "post": {
                "description": "Alternative login endpoint that accepts multiple content types",
                "consumes": [
                    "multipart/form-data",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Universal login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "$ref": "#/definitions/controllers.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to generate token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "controllers.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "password123"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "controllers.LoginResponse": {
            "type": "object",
            "properties": {
                "level": {
                    "$ref": "#/definitions/models.Level"
                },
                "level_kode": {
                    "type": "string",
                    "example": "ADM"
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "user": {
                    "$ref": "#/definitions/controllers.UserResponse"
                }
            }
        },
        "controllers.LogoutResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Logout successful"
                }
            }
        },
        "controllers.ProfileResponse": {
            "type": "object",
            "properties": {
                "level": {
                    "$ref": "#/definitions/models.Level"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                }
            }
        },
        "controllers.RefreshTokenResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Token refreshed successfully"
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
...
        "controllers.UserResponse": {
            "type": "object",
            "properties": {
                "level": {
                    "$ref": "#/definitions/models.Level"
                },
                "level_id": {
                    "type": "integer",
                    "example": 1
                },
                "user_alamat": {
                    "type": "string",
                    "example": "Jl. Contoh No. 123"
                },
                "user_email": {
                    "type": "string",
                    "example": "admin@example.com"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                },
                "user_nama": {
                    "type": "string",
                    "example": "Administrator"
                },
                "user_no_telp": {
                    "type": "string",
                    "example": "081234567890"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                },
                "warga": {
                    "description": "hanya tampil jika tidak nil",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Warga"
                        }
                    ]
                }
            }
        },
        "models.Agama": {
            "type": "object",
            "properties": {
                "agama_id": {
                    "type": "integer"
                },
                "agama_nama": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "wargas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Warga"
                    }
                }
            }
        },
        "models.KategoriProduk": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "kategori_produk_id": {
                    "type": "integer"
                },
                "kategori_produk_nama": {
                    "type": "string"
                },
                "produks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Produk"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Keluarga": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "keluarga_id": {
                    "type": "integer"
                },
                "keluarga_nama": {
                    "type": "string"
                },
                "keluarga_status": {
                    "type": "string"
                },
                "mutasiKeluargas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MutasiKeluarga"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "wargas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Warga"
                    }
                }
            }
        },
        "models.Level": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "level_id": {
                    "type": "integer"
                },
                "level_kode": {
                    "type": "string"
                },
                "level_nama": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.User"
                    }
                }
            }
        },
        "models.MutasiKeluarga": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "keluarga": {
                    "$ref": "#/definitions/models.Keluarga"
                },
                "keluarga_id": {
                    "type": "integer"
                },
                "mutasi_keluarga_alasan": {
                    "type": "string"
                },
                "mutasi_keluarga_id": {
                    "type": "integer"
                },
                "mutasi_keluarga_jenis": {
                    "type": "string"
                },
                "mutasi_keluarga_tanggal": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Pekerjaan": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "pekerjaan_id": {
                    "type": "integer"
                },
                "pekerjaan_nama": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "wargas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Warga"
                    }
                }
            }
        },
        "models.Produk": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "kategori_produk": {
                    "$ref": "#/definitions/models.KategoriProduk"
                },
                "kategori_produk_id": {
                    "type": "integer"
                },
                "produk_deskripsi": {
                    "type": "string"
                },
                "produk_foto": {
                    "type": "string"
                },
                "produk_harga": {
                    "type": "number"
                },
                "produk_id": {
                    "type": "integer"
                },
                "produk_nama": {
                    "type": "string"
                },
                "produk_stok": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.Rumah": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "rumah_alamat": {
                    "type": "string"
                },
                "rumah_id": {
                    "type": "integer"
                },
                "rumah_status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "wargas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Warga"
                    }
                }
            }
        },
        "models.TagihanIuran": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "tagihan_iuran": {
                    "type": "string"
                },
                "tagihan_iuran_nominal": {
                    "type": "number"
                },
                "tagihan_iuran_status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "warga": {
                    "$ref": "#/definitions/models.Warga"
                },
                "warga_id": {
                    "type": "integer"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "foto_profile": {
                    "type": "string"
                },
                "level": {
                    "$ref": "#/definitions/models.Level"
                },
                "level_id": {
                    "type": "integer"
                },
                "nama": {
                    "type": "string"
                },
                "no_telp": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "produk": {
                    "description": "one-to-many produk",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Produk"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                },
                "warga": {
                    "description": "one-to-one",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Warga"
                        }
                    ]
                }
            }
        },
        "models.Warga": {
            "type": "object",
            "properties": {
                "agama": {
                    "$ref": "#/definitions/models.Agama"
                },
                "agama_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "keluarga": {
                    "description": "Relasi many-to-one",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Keluarga"
                        }
                    ]
                },
                "keluarga_id": {
                    "type": "integer"
                },
                "pekerjaan": {
                    "$ref": "#/definitions/models.Pekerjaan"
                },
                "pekerjaan_id": {
                    "type": "integer"
                },
                "rumah": {
                    "$ref": "#/definitions/models.Rumah"
                },
                "rumah_id": {
                    "type": "integer"
                },
                "tagihan_iurans": {
                    "description": "Relasi one-to-many",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TagihanIuran"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "user_id": {
                    "type": "integer"
                },
                "warga_id": {
                    "type": "integer"
                },
                "warga_jenis_kelamin": {
                    "type": "string"
                },
                "warga_nama": {
                    "type": "string"
                },
                "warga_nik": {
                    "type": "string"
                },
                "warga_no_tlp": {
                    "type": "string"
                },
                "warga_status_aktif": {
                    "type": "string"
                },
                "warga_status_hidup": {
                    "type": "string"
                },
                "warga_tanggal_lahir": {
                    "type": "string"
                },
                "warga_tempat_lahir": {
                    "type": "string"
                }
            }
        }
    }
}